function/requirements.txt
google-cloud-firestore>=2.5.0
